refactor(vta): extract backup-blob byte transport to operations (P2.4 remainder-b)
Move the token-gated GET/POST backup-blob state machine out of the route
layer into operations::backup::blob (read_export_blob / write_import_blob),
a sibling to the existing operations::backup::descriptors op layer.

The state machine — token verification, export-ready / import-pending
guards, expiry, the delete-before-flip one-shot read, the size + SHA-256
upload verification, and stage-to-disk (700/600 perms) — plus the
enforce_* / gone / sha256_hex helpers now live in operations. The route
file keeps only transport concerns: bundle-id path parse, x-backup-token
header extraction, the body-cap read, and Response framing. backup_blob.rs
drops from 669 to ~260 lines; the two handlers go from ~85/70 to ~22/26.

Wire behaviour is byte-identical — the AppError variants returned are
unchanged, so the HTTP status table is preserved. The 11 router
integration tests pass unchanged as the safety net.

fmt + clippy (default & tee) clean; vta-service lib 719 + vault wire 12
green; vta-enclave checks. No version bump (additive module).
